Fix GQA api detection on PyTorch<2.5. (#438)
Revert the detection of use_gqa to before #330.
This PR bypassed the detection to enable support for multi-gpu training, and instead always enabled the use_gqa flag. However, this flag is not available if Pytorch is older than 2.5. I reverted the change to the detection by copying the previous implementation from
https://github.com/PriorLabs/TabPFN/blob/17961ba58d6d1c9ec86d0e692909ba306f76d935/src/tabpfn/model/multi_head_attention.py.

The bug was reported in #418.

In a follow-up PR I will fix multi-gpu training again.